third_party/blink/web_tests/platform/generic/fast/dom/shadow/access-key-expected.txt

Tests to ensure that accesskey works in regard to shadow DOM boundary. Can only run within DRT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dispatchedEvent("focus") is ["inputE"]
PASS dispatchedEvent("focus") is ["inputE"]
PASS dispatchedEvent("focus") is ["inputE"]
PASS successfullyParsed is true

TEST COMPLETE
